backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
motor>=3.2.0
python-jose>=3.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.1
python-dotenv>=1.0.0
google-generativeai>=0.2.0
# midnight-sdk  # Add correct package name when available